Add files via upload
Update two modules. One is to evaluate the dot product between the relative position vector and the relative velocity vector to determine the relative kinematic type. The other is to evaluate the dot product between the Euler pole and the position vector of the reference GDU or SuperGDU. Please use this evaluate_tectonic_motion.py in place of the Python with the same name in the zip folder containing all Python scripts that are products of Chapter 4 - kinematic evaluation.